Extract food selection logic from StockDistribution
